Begin extracting Superagent / transport-enabled build
