+++ superagent/tests/.eslintrc.js
@@ -1,0 +1,5 @@
+module.exports = {
+	'env': {
+		jest: true,
+	},
+};
