+++ tests/integration/posts.js
@@ -2,7 +2,7 @@
 
 const path = require( 'path' );
 
-const WPAPI = require( '../../' );
+const WPAPI = require( '../../superagent' );
 const WPRequest = require( '../../lib/constructors/wp-request.js' );
 
 // Inspecting the titles of the returned posts arrays is an easy way to
